% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf.alignment.stats.R
\name{hf.alignment.stats}
\alias{hf.alignment.stats}
\title{Calculate population genetics statistics from a PopGenome GENOME object}
\usage{
hf.alignment.stats(x, slide = F, width = 1000, step = 500)
}
\description{
Takes in a PopGenome GENOME object (from PopGenome::readData) and
calculates basic population genetics statistics. Uses hfufs if needed.
}
