% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfufs.R
\name{hfufs}
\alias{hfufs}
\title{Calculate Fu's Fs for arbitrary parameters}
\usage{
hfufs(n, k, theta)
}
\arguments{
\item{n}{The number of total sequences/individuals}

\item{k}{The number of unique alleles}

\item{theta}{The average pairwise nucleotide divergence}
}
\value{
Fu's Fs
}
\description{
Returns Fu's Fs statistic.
}
\details{
Fu's Fs is a population genetics statistic that is useful for detecting
loci that are responsible for population expansion, for example. Fu's Fs
can be formulated as a calculation that involves Stirling numbers of the
first kind. These can get large very quickly and exceed the floating point
range for modern genomic data sets.

`hfufs` calculates Fu's Fs either directly, if the numbers are small
enough, or using a logarithmic estimator for Stirling numbers implemented
in the lstirling function.
}
\examples{
n <- 100
k <- 30
theta <- 12.345
hfufs(n, k, theta)
# -0.7374915

}
\references{
Fu, Y.X. (1997) `Statistical Tests of Neutrality of Mutations Against
Population Growth, Hitchhiking and Background Selection`. Genetics
147:915-25
}
\seealso{
`hstrobecks` for a Strobeck's S calculator and `lstirling` for
the Stirling number estimator
}
