% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lstirling.R
\name{lstirling}
\alias{lstirling}
\title{Calculate the logarithm of a Stirling number of the first kind}
\usage{
lstirling(n, m)
}
\arguments{
\item{n}{The first parameter (subscript in the notation used in Temme 1993)}

\item{m}{The second parameter (superscript in parentheses in the notation
used in Temme 1993)}
}
\value{
The logarithm of the Stirling number of the first kind.
}
\description{
Returns an approximation of the logarithm of a Stirling number of the first
kind. This is useful because for larger parameter values, these Stirling
numbers can overflow the floating point range.
}
\examples{
n <- 10
m <- 5
exp(lstirling(n, m))
# 267854.5

}
\references{
Temme, N.M. (1993) Studies in Applied Mathematics 89:233-243 (Equation 3.5)
}
\seealso{
`stirmat` for a direct Stirling number calculator
}
