% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hf.readData.R
\name{hf.readData}
\alias{hf.readData}
\title{Simple interface to read an aligned fasta file into PopGenome}
\usage{
hf.readData(fasta_file)
}
\description{
Reads a given fasta file into a PopGenome GENOME object.
}
\examples{
\dontrun {
fasta_file <- "name_of_aligned.fasta"
hf.readData(fasta_file)
}

}
